chat-app/services/chat-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-jose
websockets
passlib 
sqladmin